refactor: split common into smaller chunks
